新添加了一个类专门用于生成伪彩图像；Multi窗体的Refresh与RefreshImage分离开使用，后者仅用于刷新图像；Image控件删除Refresh函数，仅保留RefreshPseudoColor函数来刷新图像。

+++ Spectra/Classes/Shared.cs
@@ -315,7 +315,7 @@ namespace Spectra
         /// </summary>
         /// <param name="wave">波长</param>
         /// <returns>谱段</returns>
-        public static int getBand(double wave)
+        public static UInt16 getBand(double wave)
         {
             double min = 2000;
             int index = 0;
@@ -327,7 +327,7 @@ namespace Spectra
                     min = wave - Convert.ToDouble(dtBandWave.Rows[i][1]);
                 }
             }
-            return Convert.ToInt32(dtBandWave.Rows[index][0]);
+            return Convert.ToUInt16(dtBandWave.Rows[index][0]);
         }
 
         /// <summary>
